requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
aiogram>=3.12.0
asyncpg>=0.29.0